+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Highlights are marked with a pancake 🥞
 - Add `TopicStore` trait and SQLite implementation [#1011](https://github.com/p2panda/p2panda/pull/1011)
 - `Ingest` processor to insert operations and associate them with topic in store [#1044](https://github.com/p2panda/p2panda/pull/1044)
 - `as_bytes` method for `Body` [#1044](https://github.com/p2panda/p2panda/pull/1044)
+- Event processing pipeline in Node stream [#1045](https://github.com/p2panda/p2panda/pull/1045)
 
 ### Changed